web/pyproject.toml
[project]
name = "web"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bilibili-api-python>=17.4.0",
    "bs4>=0.0.2",
    "ez-agent==0.2.2",
    "fastapi>=0.117.1",
    "psutil>=7.1.3",
    "uvicorn>=0.37.0",
    "volcengine-python-sdk[ark]>=4.0.31",
    "websockets>=15.0.1",
]
